requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
openai==1.47.0
weasyprint==62.3
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2

